+++ apps/calculation/additional_results/rational_list_controller.h
@@ -16,8 +16,8 @@ class RationalListController : public ChainableExpressionsListController {
 
  private:
   I18n::Message messageAtIndex(int index) override;
-  int textAtIndex(char* buffer, size_t bufferSize, Escher::HighlightCell* cell,
-                  int index) override;
+  size_t textAtIndex(char* buffer, size_t bufferSize,
+                     Escher::HighlightCell* cell, int index) override;
 };
 
 }  // namespace Calculation
