+++ apps/inference/statistic/chi_square_and_slope/goodness_table_cell.cpp
@@ -47,7 +47,7 @@ bool GoodnessTableCell::recomputeDimensionsAndReload(bool forceReloadTableCell,
   return false;
 }
 
-int GoodnessTableCell::fillColumnName(int column, char* buffer) {
+size_t GoodnessTableCell::fillColumnName(int column, char* buffer) {
   return strlcpy(buffer, I18n::translate(k_columnHeaders[column]),
                  Shared::ColumnParameterController::k_titleBufferSize);
 }
