+++ apps/shared/function.h
@@ -48,12 +48,13 @@ class Function : public ExpressionModelHandle {
   virtual float rangeStep() const { return NAN; }
 
   // Name
-  int name(char* buffer, size_t bufferSize) const;
-  virtual int nameWithArgument(char* buffer, size_t bufferSize);
+  size_t name(char* buffer, size_t bufferSize) const;
+  virtual size_t nameWithArgument(char* buffer, size_t bufferSize);
   // Insert the value of the evaluation (or the symbol if symbolValue) in buffer
-  virtual int printValue(double cursorT, double cursorX, double cursorY,
-                         char* buffer, int bufferSize, int precision,
-                         Poincare::Context* context, bool symbolValue = false);
+  virtual size_t printValue(double cursorT, double cursorX, double cursorY,
+                            char* buffer, size_t bufferSize, int precision,
+                            Poincare::Context* context,
+                            bool symbolValue = false);
   virtual I18n::Message parameterMessageName() const = 0;
 
   // Evaluation
